[IN-PROGRESS] Responsive Navbar (#11)
I made progress on step 3 of Issue #6. I added a navbar to the main three pages (home, deck, and filter) with basic styling, also setting up the structure of these files in the project directory to be able to do this.

This is still in progress since I need to make it responsive on mobile, but I wanted to prioritize getting this merged early to avoid messy merge conflicts.